requirements-mwaa.txt
# Minimal Airflow runtime needed for external_python task
apache-airflow==2.10.3
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-mysql==5.7.3

# Core iceberg dependency with S3 support & Glue catalog
pyiceberg[s3fs,glue]==0.10.0

pyarrow==19.0.1

# Needed for error capture
wurlitzer==3.1.1

# Needed for observability
opentelemetry-sdk==1.33.1
opentelemetry-exporter-otlp==1.33.1
